Add files via upload
This program simulates the process of cutting a circle out of a disk using a laser or drilling the disc. It uses the OpenCV library to create a black disk image, then a red circle is drawn on the disk to simulate the area to be cut. Finally, the result is displayed and saved to a file.